requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.2
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.13.1
boto3>=1.26.137
python-dotenv>=1.0.0
pyautogui>=0.9.54
Pillow>=9.5.0
reportlab>=4.0.4
gunicorn==21.2.0
python-jose==3.3.0
requests==2.31.0
jmespath==1.0.1
s3transfer==0.10.0
urllib3==2.2.1
PyJWT==2.8.0